+++ src/main/kotlin/dev/ekvedaras/hyperfquery/reference/BindingsReferenceContributor.kt
@@ -1,0 +1,15 @@
+package dev.ekvedaras.hyperfquery.reference
+
+import com.intellij.patterns.PlatformPatterns
+import com.intellij.psi.PsiReferenceContributor
+import com.intellij.psi.PsiReferenceRegistrar
+import com.jetbrains.php.lang.psi.elements.impl.StringLiteralExpressionImpl
+
+class BindingsReferenceContributor : PsiReferenceContributor() {
+    override fun registerReferenceProviders(registrar: PsiReferenceRegistrar) {
+        registrar.registerReferenceProvider(
+            PlatformPatterns.psiElement(StringLiteralExpressionImpl::class.java),
+            BindingsReferenceProvider()
+        )
+    }
+}